backport: merge bitcoin#20788, #19203, #19415, #21444, #21328, #20457, #20429, #18416, #16251, partial #18638 (auxiliary backports: part 7)

+++ src/net.h
@@ -761,7 +761,7 @@ extern std::string strSubVersion;
 
 struct LocalServiceInfo {
     int nScore;
-    int nPort;
+    uint16_t nPort;
 };
 
 extern CCriticalSection cs_mapLocalHost;
